+++ src/prim/point_2dh.inl
@@ -85,7 +85,7 @@ Point2DH<T>::position() const
 
 template<typename T> inline
 typename Point2DH<T>::TConstReference
-Point2DH<T>::operator[](unsigned iIndex) const
+Point2DH<T>::operator[](size_t iIndex) const
 {
 	LASS_ASSERT(iIndex < dimension);
 	return *(&x + iIndex);
@@ -95,7 +95,7 @@ Point2DH<T>::operator[](unsigned iIndex) const
 
 template<typename T> inline
 typename Point2DH<T>::TReference
-Point2DH<T>::operator[](unsigned iIndex)
+Point2DH<T>::operator[](size_t iIndex)
 {
 	LASS_ASSERT(iIndex < dimension);
 	return *(&x + iIndex);
