+++ src/prim/simple_polygon_2d.h
@@ -81,8 +81,8 @@ public:
 	template <typename InputIterator> 
 	SimplePolygon2D(InputIterator iFirstVertex, InputIterator iLastVertex);
 
-	const TPoint& operator[](int iIndexOfVertex) const;
-	TPoint& operator[](int iIndexOfVertex);
+	const TPoint& operator[](size_t iIndexOfVertex) const;
+	TPoint& operator[](size_t iIndexOfVertex);
 	const TPoint& at(int iIndexOfVertex) const;
 	TPoint& at(int iIndexOfVertex);
 	const TLineSegment edge(int iIndexOfTailVertex) const;
@@ -93,7 +93,7 @@ public:
 	void erase(int iIndexOfVertex);
 
 	const bool isEmpty() const;
-	const int size() const;
+	const size_t size() const;
 
 	const TValue signedArea() const;
 	const TValue area() const;
